risk-engine/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.0
numpy
scikit-learn
xgboost
shap
pandas
python-dotenv
httpx
